requirements.txt
# --- Core Server ---
fastapi==0.115.12
uvicorn==0.34.3
python-dotenv==1.1.0
weasyprint==61.2
jinja2==3.1.3
openai>=1.0.0
langchain>=0.1.13
langchain-community>=0.0.24

# --- Database ---
SQLAlchemy==2.0.41
psycopg2-binary==2.9.10
asyncpg==0.30.0
alembic==1.16.2

# --- Auth & Security ---
passlib==1.7.4
bcrypt==4.1.2
python-jose==3.5.0
python-multipart==0.0.20

# --- Pydantic (v2) ---
pydantic==2.11.5
pydantic_core==2.33.2
typing_extensions==4.14.0
annotated-types==0.7.0

# --- Optional: FastAPI Users if actively used ---
fastapi-users==14.0.1
email_validator==2.2.0

# --- Dev & Testing Tools (optional but useful) ---
black==25.1.0
ruff==0.11.13
pytest==8.4.0
pytest-cov==6.2.1
coverage==7.9.1
requests

